# E-commerceApp.Api/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "E-commerceApp.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Brands/{id}": {
      "get": {
        "tags": [
          "Brands"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BrandGetDto"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BrandPutDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Brands": {
      "post": {
        "tags": [
          "Brands"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BrandPostDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Brands"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BrandGetPaginatedDtoPaginatedListDto"
            }
          }
        }
      }
    },
...
    "/api/Sliders": {
      "post": {
        "tags": [
          "Sliders"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Title",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "SecondTitle",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Description",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ButtonUrl",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Order",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "ButtonText",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "BgColor",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ImageFile",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Sliders/{id}": {
      "get": {
        "tags": [
          "Sliders"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SliderGetDto"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sliders"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Title",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "SecondTitle",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Description",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ButtonUrl",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ButtonText",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Order",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "BgColor",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ImageFile",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Sliders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BrandGetDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "productsCount": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "BrandGetPaginatedDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "productsCount": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "BrandGetPaginatedDtoPaginatedListDto": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BrandGetPaginatedDto"
          }
        },
        "pageIndex": {
          "format": "int32",
          "type": "integer"
        },
        "totalPages": {
          "format": "int32",
          "type": "integer"
        },
        "havNext": {
          "type": "boolean",
          "readOnly": true
        },
        "havPrev": {
          "type": "boolean",
          "readOnly": true
        }
      }
    },
...
    "BrandPostDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "BrandPutDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
...
    "SliderGetDto": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "secondTitle": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "buttonUrl": {
          "type": "string"
        },
        "buttonText": {
          "type": "string"
        },
        "imageUri": {
          "type": "string"
        },
        "bgColor": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}