# E-commerceApp.Service/Dtos/SliderDtos/SliderGetDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace E_commerceApp.Service.Dtos.SliderDtos
{
    public class SliderGetDto
    {
        public string Title { get; set; }

        public string SecondTitle { get; set; }

        public string Description { get; set; }

        public string ButtonUrl { get; set; }

        public string ButtonText { get; set; }

        public string ImageUri { get; set; }

        public string BgColor { get; set; }
    }
}